refactor: refactoring structure
